+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/EyesAppiumUtils.java
@@ -43,25 +43,6 @@ public class EyesAppiumUtils {
         return platformVersionObj == null ? null : String.valueOf(platformVersionObj);
     }
 
-    /**
-     * @param driver The driver to get the platform version from.
-     * @return The device name or 'Unknown' if it is undefined.
-     */
-    public static String getDeviceName(HasCapabilities driver) {
-        Capabilities capabilities = driver.getCapabilities();
-
-        Object deviceNameCapability = capabilities.getCapability("deviceName");
-        String deviceName = deviceNameCapability != null ? deviceNameCapability.toString() : "Unknown";
-
-        Object deviceCapability = capabilities.getCapability("device");
-
-        if (deviceCapability != null && !deviceName.toLowerCase().contains(deviceCapability.toString())) {
-            deviceName = deviceCapability.toString();
-        }
-
-        return deviceName;
-    }
-
     public static WebElement getFirstScrollableView(WebDriver driver) {
         return driver.findElement(By.xpath(SCROLLVIEW_XPATH));
     }