+++ eyes.sdk.core/src/main/java/com/applitools/eyes/config/Feature.java
@@ -2,5 +2,6 @@ package com.applitools.eyes.config;
 
 public enum Feature {
     UNUSED,
-    NO_SWITCH_WITHOUT_FRAME_CHAIN
+    NO_SWITCH_WITHOUT_FRAME_CHAIN,
+    OPTIMIZE_MOBILE_DEVICES_INFO
 }