+++ eyes.selenium.java/src/test/java/com/applitools/eyes/selenium/MobileNativeTests.java
@@ -1,5 +1,7 @@
 package com.applitools.eyes.selenium;
 
+import com.applitools.eyes.config.Configuration;
+import com.applitools.eyes.config.Feature;
 import com.applitools.eyes.selenium.fluent.Target;
 import com.applitools.eyes.utils.ReportingTestSuite;
 import com.applitools.eyes.utils.SeleniumTestUtils;
@@ -7,8 +9,10 @@ import io.appium.java_client.MobileBy;
 import io.appium.java_client.MobileElement;
 import io.appium.java_client.android.AndroidDriver;
 import io.appium.java_client.android.AndroidElement;
+import io.appium.java_client.ios.IOSDriver;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.net.MalformedURLException;
@@ -22,7 +26,7 @@ public class MobileNativeTests extends ReportingTestSuite {
         super.setGroupName("selenium");
     }
 
-    private void setCapabilities(Eyes eyes, DesiredCapabilities capabilities, String methodName) {
+    private void setCapabilities(DesiredCapabilities capabilities, String methodName) {
         capabilities.setCapability("username", SAUCE_USERNAME);
         capabilities.setCapability("accesskey", SAUCE_ACCESS_KEY);
         capabilities.setCapability("name", methodName);
@@ -32,7 +36,7 @@ public class MobileNativeTests extends ReportingTestSuite {
         Eyes eyes = new Eyes();
         String testName = Thread.currentThread().getStackTrace()[2].getMethodName();
         SeleniumTestUtils.setupLogging(eyes, testName);
-        setCapabilities(eyes, capabilities, testName);
+        setCapabilities(capabilities, testName);
         eyes.setBatch(TestDataProvider.batchInfo);
         eyes.setSaveNewTests(false);
         return eyes;
@@ -98,35 +102,56 @@ public class MobileNativeTests extends ReportingTestSuite {
         }
     }
 
-//    @Test
-//    public void iOSNativeAppTest() throws Exception {
-//
-//        DesiredCapabilities caps = DesiredCapabilities.iphone();
-//        caps.setCapability("appiumVersion", "1.13.0");
-//        caps.setCapability("deviceName", "iPhone XS Simulator");
-//        caps.setCapability("deviceOrientation", "portrait");
-//        caps.setCapability("platformVersion", "12.2");
-//        caps.setCapability("platformName", "iOS");
-//        caps.setCapability("browserName", "");
-//        caps.setCapability("app", "https://applitools.bintray.com/Examples/HelloWorldiOS_1_0.zip");
-//
-//        Eyes eyes = initEyes(caps);
-//        eyes.setLogHandler(new StdoutLogHandler());
-//        eyes.setBatch(TestDataProvider.batchInfo);
-//
-//
-//        WebDriver driver = null;
-//        try {
-//            driver = new IOSDriver(new URL(SAUCE_SELENIUM_URL), caps);
-//            eyes.open(driver, "Mobile Native Tests", "iOS Native App");
-//            eyes.checkWindow("checkWindow");
-//            eyes.close();
-//        } finally {
-//            if (driver != null) {
-//                driver.quit();
-//            }
-//
-//            eyes.abortIfNotClosed();
-//        }
-//    }
+    @Test
+    public void TestOptimizeMobileInfoFeatureAndroid() throws MalformedURLException {
+        DesiredCapabilities capabilities = new DesiredCapabilities();
+        capabilities.setCapability("platformName", "Android");
+        capabilities.setCapability("deviceName", "Samsung Galaxy S7 GoogleAPI Emulator");
+        capabilities.setCapability("platformVersion", "8.0");
+        capabilities.setCapability("app", "http://saucelabs.com/example_files/ContactManager.apk");
+        capabilities.setCapability("clearSystemFiles", true);
+        capabilities.setCapability("noReset", true);
+
+        Eyes eyes = initEyes(capabilities);
+        Configuration configuration = eyes.getConfiguration();
+        configuration.setFeatures(Feature.OPTIMIZE_MOBILE_DEVICES_INFO);
+
+        AndroidDriver driver = new AndroidDriver(new URL(SAUCE_SELENIUM_URL), capabilities);
+        try {
+            eyes.open(driver, "Mobile Native Tests", "Test Optimize Mobile Info Feature Android");
+            eyes.checkWindow("Contact list");
+            eyes.close();
+        } finally {
+            driver.quit();
+            eyes.abortIfNotClosed();
+        }
+
+        Assert.assertEquals(eyes.getDevicePixelRatio(), 4.0);
+    }
+
+    @Test
+    public void TestOptimizeMobileInfoFeatureIos() throws Exception {
+        DesiredCapabilities caps = new DesiredCapabilities();
+        caps.setCapability("deviceName", "iPhone 8 Plus Simulator");
+        caps.setCapability("deviceOrientation", "portrait");
+        caps.setCapability("platformVersion", "13.0");
+        caps.setCapability("platformName", "iOS");
+        caps.setCapability("app", "https://applitools.bintray.com/Examples/HelloWorldiOS_1_0.zip");
+
+        Eyes eyes = initEyes(caps);
+        Configuration configuration = eyes.getConfiguration();
+        configuration.setFeatures(Feature.OPTIMIZE_MOBILE_DEVICES_INFO);
+
+        WebDriver driver = new IOSDriver(new URL(SAUCE_SELENIUM_URL), caps);
+        try {
+            eyes.open(driver, "Mobile Native Tests", "Test Optimize Mobile Info Feature Ios");
+            eyes.checkWindow();
+            eyes.close();
+        } finally {
+            driver.quit();
+            eyes.abortIfNotClosed();
+        }
+
+        Assert.assertEquals(eyes.getDevicePixelRatio(), 3.0);
+    }
 }